backend/requirements.txt
# FastAPI y servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Base de datos
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Pydantic y validación
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.1.0

# Autenticación y seguridad
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Storage
boto3>=1.28.0
cloudinary>=1.36.0

# Email
sendgrid>=6.11.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
